affichage des joueurs sur la route /team
