+++ backend/package.json
@@ -8,6 +8,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "mysql2": "^3.5.2"
