+++ backend/src/app.js
@@ -4,8 +4,6 @@ const express = require("express");
 
 const app = express();
 
-const teamControllers = require("./controllers/teamControllers");
-
 // Configure it
 
 /* ************************************************************************* */
@@ -27,20 +25,15 @@ const teamControllers = require("./controllers/teamControllers");
 // 4. Be sure to only have URLs in the array with domains from which you want to allow requests.
 // For example: ["http://mysite.com", "http://another-domain.com"]
 
-/*
 const cors = require("cors");
 
 app.use(
   cors({
-    origin: [
-      process.env.FRONTEND_URL, // keep this one, after checking the value in `backend/.env`
-      "http://mysite.com",
-      "http://another-domain.com",
-    ]
+    origin: [process.env.FRONTEND_URL],
   })
 );
-*/
 
+const teamControllers = require("./controllers/teamControllers");
 /* ************************************************************************* */
 
 // Request Parsing: Understanding the purpose of this part
