+++ frontend/src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import React from "react";
 
 export default function Home() {
-  return <h1>TEST</h1>;
+  return <h1>TEST HOME</h1>;
 }
